backend/requirements.txt
autopep8
pydicom
pynetdicom
Flask
flask_cors
pycryptodomex
cryptography
waitress
jstyleson
boto3
requests
psutil